backend/requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
Django==4.1
django-cors-headers==4.3.0
django-filter==23.3
django-restframework==0.0.1
djangorestframework==3.14.0
fiona==1.9.5
packaging==23.2
psycopg2==2.9.9
pyproj==3.6.1
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
